refactor(repl): rename groovy-repl to repl with gvy.repl package
Simplify module naming as part of broader gvy.* package standardization:

- Rename groovy-repl/ → repl/

- Package: com.github.albertocavalcante.groovyrepl → com.github.albertocavalcante.gvy.repl

- Update dependencies in jupyter:kernel-core and jupyter:kernels:groovy
