+++ settings.gradle.kts
@@ -40,7 +40,7 @@ include("groovy-junit")
 include("jupyter:kernel-core")
 include("jupyter:kernels:groovy")
 include("jupyter:kernels:jenkins")
-include("groovy-repl")
+include("repl")
 include("tools:jenkins-extractor")
 include("viz:ast-model")
 include("viz:desktop")
